refactor: rename abi_check → abicheck, abi-check → abicheck throughout
- Rename package directory abi_check/ → abicheck/
- Update all imports, CLI entrypoint, pyproject.toml, README, NOTICE, CI
- Fix all ruff lint issues in new package (UP035, UP045, F401, I001, F541)
- GitHub repo already renamed to napetrov/abicheck via API
- 34/34 tests green, ruff clean
