+++ abicheck.egg-info/SOURCES.txt
@@ -1,0 +1,20 @@
+LICENSE
+NOTICE.md
+README.md
+pyproject.toml
+abicheck/__init__.py
+abicheck/checker.py
+abicheck/cli.py
+abicheck/dumper.py
+abicheck/model.py
+abicheck/reporter.py
+abicheck/serialization.py
+abicheck.egg-info/PKG-INFO
+abicheck.egg-info/SOURCES.txt
+abicheck.egg-info/dependency_links.txt
+abicheck.egg-info/entry_points.txt
+abicheck.egg-info/requires.txt
+abicheck.egg-info/top_level.txt
+tests/test_checker.py
+tests/test_reporter.py
+tests/test_serialization.py
