+++ abicheck/reporter.py
@@ -105,6 +105,6 @@ def to_markdown(result: DiffResult) -> str:
         "| ⚠️ SOURCE_BREAK | Source-level API change, binary compatible |",
         "| ❌ BREAKING | Binary ABI break — recompilation required |",
         "",
-        "_Generated by [abi-check](https://github.com/napetrov/abi-check)_",
+        "_Generated by [abicheck](https://github.com/napetrov/abicheck)_",
     ]
     return "\n".join(lines)
